Ref: extracting new method processContainers().
